Add files via upload
https://s3.us-east-2.amazonaws.com/datafaculty/final_case/final_data.zip
This data set contains more than 500 high resolution images of common world objects. The images in this dataset belong to the following classes:
1. Acoustic guitar
2. Airliner
3. Arabian camel
4. Chihuahua
5. Church
6. German Shepherd
7. Laptop
8. Tabby
9. Violin
Your task is to use a pre-trained model such as VGG16, to arrange these images in new folders, where images of one class are in one folder.